[업그레이드] 파워, 시간 업그레이드 비지니스 로직 추가

+++ Assets/02_script/network/SHNetworkManager_Web.cs
@@ -76,7 +76,7 @@ public partial class SHNetworkManager : SHSingleton<SHNetworkManager>
 
         DebugLogOfWebResponse(pReply);
 
-        if (true == (m_bIsConnectWebServer = (eErrorCode.Net_Common_HTTP != pReply.errorCode)))
+        if (true == (m_bIsConnectWebServer = ((int)eErrorCode.Net_Common_HTTP != pReply.errorCode)))
         {
             pRequestData.m_pCallback(pReply);
             m_pWebRequestQueue.Remove(pRequestData);
@@ -139,7 +139,7 @@ public partial class SHNetworkManager : SHSingleton<SHNetworkManager>
 
         DebugLogOfWebResponse(pReply);
 
-        if (true == (m_bIsConnectWebServer = (eErrorCode.Net_Common_HTTP != pReply.errorCode)))
+        if (true == (m_bIsConnectWebServer = ((int)eErrorCode.Net_Common_HTTP != pReply.errorCode)))
         {
             StopRetryProcess();
